made the high level functions part of a class
